packages/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.26.0
plaid-python>=15.0.0
psycopg2-binary>=2.9.7
alembic>=1.12.0
SQLAlchemy>=2.0.23
redis>=5.0.0
celery>=5.3.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
python-json-logger>=2.0.7
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.7.0
# Security updates - Mon Sep  8 19:00:54 PDT 2025
urllib3>=2.5.0  # Fix SSRF vulnerability
# ML dependencies
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
